changed example service to 16 bit value

+++ TankOS/twi/services/example.c
@@ -4,6 +4,6 @@
 
 int test_read_value_format(int (*print)(const char *fmt, ...), void *results, uint16_t results_length) {
     if (results_length != sizeof(int32_t)) return 0;
-    int32_t *res = (int32_t*) results;
-    return print("Test value: %u", *res);
+    int16_t *res = (int16_t*) results;
+    return print("Test value: %i", (int) *res);
 }
